Expr: the "id" is now a universal id (not for the function the node belongs to) set at the construction of the node NodeMap: use "id" instead of pointer of ExprNode* as hash key ---> reproductibility Domain: minor comment change HC4Revise: comment out proj(const Function& f, const Domain& y, Array<Domain>& x)

+++ src/symbolic/ibex_Expr.h
@@ -71,10 +71,13 @@ public:
 	 * two subnodes referencing the same object count for 1).*/
 	const int size;
 
-	/** Unique number identifying this node in the global function.
+	/** Unique number identifying this expression node
 	 *
-	 * Undefined while this node is not part of a function expression. */
-	const int id;
+	 * This field is typically used as key in hash maps. The physical address of the
+	 * object is indeed not a good candidate (for being a key) because reproductibility
+	 * is lost in this case.
+	 */
+	const long id;
 
 	/** Dimensions */
 	const Dim dim;
